PHV allocation now respects deparser packing constraint (#2503)
* PHV allocation now respects deparser no-repeat PHV constraint

Along with this constraint, there are other code related changes in this PR:
* Remove unused classes from the repository
* Reorganize the SymBitMatrix objects maintained by PhvInfo and provide proper accessors to each of
  those matrices.

P4C-813 #assign #start-work #fixed PHV allocation now respects deparser no-repeat PHV constraint #fixed

* Remove two more lines of now dead code

P4C-813 #assign #start-work #fixed Remove two more lines of now dead code #fixed

+++ bf-p4c/phv/pragma/pa_container_size.cpp
@@ -375,7 +375,7 @@ void PragmaContainerSize::check_and_add_constraint(
     }
 
     // Ensure that the new pragma specifications are the same as the already existing ones.
-    for (int i = 0; i < existingSizePragmas.size(); i++) {
+    for (size_t i = 0; i < existingSizePragmas.size(); i++) {
         if (sizes[i] != existingSizePragmas[i]) {
             ::error("@pragma pa_container_size already specifies containers %1% for field %2%."
                     " However, to fit within learning quanta size constraints, the field must "
